Modularize cover rendering into per-template classes
