+++ src/features/epub-maker/domain/cover/templates/template-utils.ts
@@ -1,0 +1,36 @@
+import type { CoverCanvasMetrics } from "../core-types";
+
+export function createScaleHelpers(metrics: CoverCanvasMetrics) {
+  return {
+    sx: (value: number) => value * metrics.scaleX,
+    sy: (value: number) => value * metrics.scaleY,
+    su: (value: number) => value * metrics.unitScale,
+  };
+}
+
+export function drawRoundedRectPath(
+  context: CanvasRenderingContext2D,
+  x: number,
+  y: number,
+  width: number,
+  height: number,
+  radius: number,
+) {
+  const clampedRadius = Math.max(0, Math.min(radius, width / 2, height / 2));
+  context.beginPath();
+  context.moveTo(x + clampedRadius, y);
+  context.lineTo(x + width - clampedRadius, y);
+  context.quadraticCurveTo(x + width, y, x + width, y + clampedRadius);
+  context.lineTo(x + width, y + height - clampedRadius);
+  context.quadraticCurveTo(
+    x + width,
+    y + height,
+    x + width - clampedRadius,
+    y + height,
+  );
+  context.lineTo(x + clampedRadius, y + height);
+  context.quadraticCurveTo(x, y + height, x, y + height - clampedRadius);
+  context.lineTo(x, y + clampedRadius);
+  context.quadraticCurveTo(x, y, x + clampedRadius, y);
+  context.closePath();
+}
